requirements.txt
asgiref==3.9.1
Django==5.2.6
django-cors-headers==4.4.0
django-environ==0.12.0
django-filter==24.2
django-phonenumber-field==8.1.0
djangorestframework==3.15.2
djangorestframework_simplejwt==5.5.1
drf-yasg==1.21.7
inflection==0.5.1
packaging==25.0
psycopg==3.2.0
PyJWT==2.10.1
pytz==2025.2
PyYAML==6.0.2
sqlparse==0.5.3
typing_extensions==4.15.0
uritemplate==4.2.0
